Distance matrix to data.frame

Coerces distance matrix to a data.frame object

Usage

dmatrix.df(x, rm.diag = TRUE)

Arguments

x

Symmetrical distance matrix

rm.diag

(TRUE/FALSE) remove matrix diagonal, self values.

Value

data.frame object representing to and from values

Note

Function results in data.frame object with "X1" (FROM), "X2" (TO) and "distance" columns. The FROM column represents to origin ID, TO represents destination ID and distance is the associated matrix distance. These results can be joined back to the graph object using either the origin or destination ID's.

Author

Jeffrey S. Evans <jeffrey_evans@tnc.org> and Melanie Murphy <melanie.murphy@uwyo.edu>

Examples

 library(sp)
 pts <- cbind(x=runif(15, 480933, 504250), y=runif(15, 4479433, 4535122))
   pts <- SpatialPointsDataFrame(pts, 
            data.frame(ID=paste("ob",1:nrow(pts),sep="")))
 
 # Create distance matrix  
 dm <- spDists(pts, pts)  
   colnames(dm) <- pts@data[,"ID"] 
   rownames(dm) <- pts@data[,"ID"]
 
 # Coerce to data.frame with TO and FROM ID's and associated distance
 dm.df <- dmatrix.df(dm)
   head(dm.df)
#>    X1  X2  distance
#> 2 ob2 ob1 41956.941
#> 3 ob3 ob1  9321.240
#> 4 ob4 ob1 41798.056
#> 5 ob5 ob1  8060.958
#> 6 ob6 ob1  4955.728
#> 7 ob7 ob1 36720.629