Statistics for edges (lines)

Extracts raster values for each edge and calculates specified statistics

Usage

graph.statistics(x, r, stats = c("min", "mean", "max"), buffer = NULL)

Arguments

x

sp SpatialLinesDataFrame or sf LINE object

r

A terra SpatRast or raster rasterLayer, rasterStack, rasterBrick object

stats

Statistics to calculate. If vectorized, can pass a custom statistic function.

buffer

Buffer distance, radius in projection units. For statistics based on edge buffer distance

Value

data.frame object of statistics

Note

If the buffer argument is specified that, raster values within the specified buffer radius are extracted and included in the derived statistic(s). Else-wise, the statistics are derived from raster values that directly intersect each edge.

Examples

# \donttest{
 library(sp)
 library(sf)
#> Linking to GEOS 3.9.1, GDAL 3.2.1, PROJ 7.2.1; sf_use_s2() is TRUE
 library(raster)
#> 
#> Attaching package: 'raster'
#> The following object is masked from 'package:nlme':
#> 
#>     getData
 library(terra)  
#> terra 1.5.17
   data(rasters)
   data(ralu.site)

 xvars <- rast(stack(rasters))
 ralu.site <- as(ralu.site, "sf")
 
  dist.graph <- knn.graph(ralu.site, row.names = ralu.site$SiteName, 
                          max.dist = 1500)
             
   str(dist.graph@data)
#> Error in str(dist.graph@data): trying to get slot "data" from an object (class "sf") that is not an S4 object 
   
 skew <- function(x, na.rm = TRUE) {  
           if (na.rm) x <- x[!is.na(x)]
           sum( (x - mean(x)) ^ 3) / ( length(x) * sd(x) ^ 3 )  
     }

# Moments on continuous raster data
 system.time( {    
  stats <- graph.statistics(dist.graph, r = xvars[[-6]],  
              stats = c("min", "median", "max", "var", "skew")) 
 } ) 
#> Error in get(as.character(FUN), mode = "function", envir = envir): object 'skew' of mode 'function' was not found
#> Timing stopped at: 0.11 0 0.13

# Proportional function on nominal raster data    
p <- function(x) { length(x[x < 52]) / length(x) }  
  
  system.time( {    
   nstats <- graph.statistics(dist.graph, r = xvars[[6]],
               stats = "p") 
  } )   
#> Error in get(as.character(FUN), mode = "function", envir = envir): object 'p' of mode 'function' was not found
#> Timing stopped at: 0.15 0 0.16

# Based on 500m buffer distance around line(s)
 system.time( {    
  stats <- graph.statistics(dist.graph, r = xvars[[-6]],  
              stats = c("min", "median", "max", "var", "skew"),
            buffer = 500) 
 } )
#> Using 500 distance for statistics
#> Error in get(as.character(FUN), mode = "function", envir = envir): object 'skew' of mode 'function' was not found
#> Timing stopped at: 0.45 0 0.48

 dist.graph@data <- data.frame(dist.graph@data, stats, nstats)
#> Error in data.frame(dist.graph@data, stats, nstats): trying to get slot "data" from an object (class "sf") that is not an S4 object 
   str(dist.graph@data)
#> Error in str(dist.graph@data): trying to get slot "data" from an object (class "sf") that is not an S4 object 
# }