Effect Size

Cohen's D effect size for gravity models

Usage

gravity.es(x, actual.n = FALSE, alpha = 0.95)

Arguments

x

gravity model object

actual.n

(FALSE/TRUE) Use actual N or degrees of freedom in calculating Confidence Interval

alpha

confidence interval

Value

data.frame of parameter effect size

Details

Calculate Cohen's D statistic for each effect in a gravity model object

References

Murphy M.A., R. Dezzani, D.S. Pilliod & A.S. Storfer (2010) Landscape genetics of high mountain frog metapopulations. Molecular Ecology 19(17):3634-3649

Cohen, J. (1988) Statistical power for the behavioral sciences (2nd ed.). Hillsdale, NJ: Erlbaum

Author

Jeffrey S. Evans <jeffrey_evans@tnc.org> and Melanie Murphy <melanie.murphy@uwyo.edu>

Examples

 
library(nlme)
  data(ralu.model)

x = c("DEPTH_F", "HLI_F", "CTI_F", "cti", "ffp")
gm_h1 <- gravity(y = "DPS", x = x, d = "DISTANCE", group = "FROM_SITE", 
                data = ralu.model, ln = FALSE, method="ML") 
#> [1] "Running singly-constrained gravity model"

gravity.es(gm_h1)
#>            t.value  df    cohen.d  p.value     low.ci      up.ci
#> DEPTH_F   2.957913  15  1.5274596 0.009775  0.6404823  2.4144368
#> HLI_F     1.243231  15  0.6420016 0.232865 -0.1533054  1.4373086
#> CTI_F    -1.193421  15 -0.6162801 0.251239 -1.4099520  0.1773917
#> cti      -2.381153 168 -0.3674198 0.018378 -0.5845083 -0.1503314
#> ffp       6.577078 168  1.0148652 0.000000  0.7860803  1.2436502
#> DISTANCE -9.258849 168 -1.4286714 0.000000 -1.6699903 -1.1873525