Predict gravity model

predict method for class "gravity"

Usage

# S3 method for gravity
predict(
  object,
  newdata,
  groups = NULL,
  back.transform = c("none", "simple", "Miller", "Naihua"),
  ...
)

Arguments

object

Object of class gravity

newdata

New data used for obtaining the predictions, can be a data.frame or nffGroupedData

groups

Grouping factor acting as random effect. If used, must match levels used in model, otherwise leave it null and do not convert to groupedData

back.transform

Method to back transform data, default is none and log predictions will be returned.

...

Arguments passed to predict.lme or predict.lm

Value

Vector of model predictions

Details

Please note that the entire gravity equation is log transformed so, your parameter space is on a log scale, not just y. This means that for a meaningful prediction the "newdata" also needs to be on a log scale.

For the back.transform argument, the simple back-transform method uses the form exp(y-hat)0.5*variance whereas Miller uses exp(sigma)*05 as the multiplicative bias factor. Naihua regresses y~exp(y-hat) with no intercept and uses the resulting coefficient as the multiplicative bias factor. The Naihua method is intended for results with non-normal errors. You can check the functional form by simply plotting y (non-transformed) against the fit. The default is to output the log scaled predictions.

References

Miller, D.M. (1984) Reducing Transformation Bias in Curve Fitting The American Statistician. 38(2):124-126

Naihua, D. (1983) Smearing Estimate: A Nonparametric Retransformation Method Journal of the American Statistical Association, 78(383):605–610.

Examples

library(nlme)
data(ralu.model)

back.transform <- function(y) exp(y + 0.5 * stats::var(y))
rmse = function(p, o){ sqrt(mean((p - o)^2)) } 

x = c("DEPTH_F", "HLI_F", "CTI_F", "cti", "ffp")
 
sidx <- sample(1:nrow(ralu.model), 100) 
  train <- ralu.model[sidx,]
  test <- ralu.model[-sidx,]
 
 # Specify constrained gravity model  
 ( gm <- gravity(y = "DPS", x = x, d = "DISTANCE", group = "FROM_SITE", 
                 data = train, ln = FALSE) )
#> [1] "Running singly-constrained gravity model"
#> Gravity model
#> 
#> Linear mixed-effects model fit by REML
#>   Data: gdata 
#>        AIC       BIC logLik
#>   -136.514 -113.7206 77.257
#> 
#> Random effects:
#>  Formula: DPS ~ 1 | FROM_SITE
#>         (Intercept)   Residual
#> StdDev:  0.04331721 0.09251116
#> 
#> Fixed effects:  list(fmla) 
#>                 Value Std.Error DF   t-value p-value
#> (Intercept) -5.021556 2.2138316 79 -2.268265  0.0260
#> DEPTH_F      0.016064 0.0142664 14  1.126013  0.2791
#> HLI_F        0.057311 0.2058492 14  0.278415  0.7848
#> CTI_F       -0.056346 0.0637005 14 -0.884544  0.3913
#> cti         -0.168146 0.0981068 79 -1.713907  0.0905
#> ffp          0.897728 0.2040685 79  4.399147  0.0000
#> DISTANCE    -0.125777 0.0154870 79 -8.121503  0.0000
#>  Correlation: 
#>          (Intr) DEPTH_ HLI_F  CTI_F  cti    ffp   
#> DEPTH_F  -0.626                                   
#> HLI_F    -0.824  0.859                            
#> CTI_F     0.330 -0.656 -0.560                     
#> cti      -0.024  0.050  0.012 -0.128              
#> ffp      -0.777  0.129  0.301  0.003 -0.138       
#> DISTANCE -0.275  0.107  0.140 -0.050  0.464  0.164
#> 
#> Standardized Within-Group Residuals:
#>         Min          Q1         Med          Q3         Max 
#> -2.96880469 -0.49096445  0.06522563  0.53632544  1.79390405 
#> 
#> Number of Observations: 100
#> Number of Groups: 18 
  
( p <- predict(gm, test[,c(x, "DISTANCE")]) )
#> Making population level predictions
#> Warning: the condition has length > 1 and only the first element will be used
#>  [1] -0.25329785 -0.25681171 -0.32777182 -0.36893756 -0.35310797 -0.48920904
#>  [7] -0.29494511 -0.31408256 -0.32889162 -0.39437858 -0.41729481 -0.47130315
#> [13] -0.44324631 -0.48485183 -0.28427044 -0.38452308 -0.39975697 -0.47981240
#> [19] -0.53065871 -0.53769501 -0.52562637 -0.58895763 -0.59249218 -0.11374265
#> [25] -0.13735936 -0.25663483 -0.44048213 -0.39306026 -0.44350423 -0.19303334
#> [31] -0.24576525 -0.46035810 -0.44332558 -0.44598750 -0.54152098 -0.21258035
#> [37] -0.22369599 -0.39416800 -0.39778379 -0.38034004 -0.38286170 -0.47873097
#> [43] -0.32367198 -0.33458788 -0.39773901 -0.41225972 -0.42571239 -0.43933051
#> [49] -0.49059890  0.04233915 -0.30908972 -0.36161362 -0.40026709 -0.41574314
#> [55] -0.12441081 -0.20873610 -0.21948389 -0.27972325 -0.31842001 -0.29592246
#> [61] -0.35275876 -0.40787969 -0.22871113 -0.28291477 -0.26784625 -0.33924191
#> [67] -0.34477744 -0.43280251 -0.04839582 -0.16087642 -0.22765633 -0.33484413
#> [73] -0.37164905 -0.23882341 -0.36709540 -0.36564847 -0.38908496 -0.17073254
#> [79] -0.39855034 -0.42117103 -0.40857146 -0.42724038 -0.37287089 -0.38934601
#> [85] -0.43209934 -0.45592882 -0.45409103 -0.40959210 -0.45613108 -0.21846319
#> attr(,"label")
#> [1] "Predicted values"
  rmse(back.transform(p), back.transform(ralu.model[,"DPS"][-sidx]))
#> [1] 0.0761695

# WIth model sigma-based back transformation
( p <- predict(gm, test[,c(x, "DISTANCE")], back.transform = "simple") )
#> Making population level predictions
#> Back-transforming exp(y-hat)*0.5*variance(y-hat), 
#> 		  assumes normally distributed errors
#>  [1] 0.7815808 0.7788393 0.7254880 0.6962291 0.7073379 0.6173322 0.7496986
#>  [8] 0.7354877 0.7246761 0.6787398 0.6633625 0.6284857 0.6463687 0.6200280
#> [15] 0.7577443 0.6854622 0.6750991 0.6231604 0.5922671 0.5881143 0.5952551
#> [22] 0.5587258 0.5567544 0.8986321 0.8776580 0.7789770 0.6481579 0.6796352
#> [29] 0.6462020 0.8301306 0.7874904 0.6354023 0.6463175 0.6445993 0.5858685
#> [36] 0.8140616 0.8050629 0.6788827 0.6764325 0.6883355 0.6866019 0.6238347
#> [43] 0.7284685 0.7205599 0.6764628 0.6667110 0.6578020 0.6489047 0.6164748
#> [50] 1.0504307 0.7391691 0.7013470 0.6747548 0.6643926 0.8890963 0.8171971
#> [57] 0.8084611 0.7611977 0.7323045 0.7489663 0.7075849 0.6696376 0.8010355
#> [64] 0.7587722 0.7702924 0.7172142 0.7132550 0.6531546 0.9593160 0.8572589
#> [71] 0.8018809 0.7203753 0.6943439 0.7929760 0.6975129 0.6985229 0.6823423
#> [78] 0.8488512 0.6759141 0.6607961 0.6691746 0.6567977 0.6934960 0.6821642
#> [85] 0.6536141 0.6382229 0.6393969 0.6684919 0.6380938 0.8092867
#> attr(,"label")
#> [1] "Predicted values"
( p <- predict(gm, test[,c(x, "DISTANCE")], back.transform = "Miller") )
#> Making population level predictions
#> Miller back-transformation using: 
#> 		  exp(sigma)*0.5*exp(y-hat)0.5*variance(y-hat), 
#> 		  assumes normally distributed errors
#>  [1] 0.8185825 0.8157111 0.7598341 0.7291901 0.7408247 0.6465580 0.7851909
#>  [8] 0.7703072 0.7589837 0.7108727 0.6947674 0.6582395 0.6769691 0.6493814
#> [15] 0.7936174 0.7179134 0.7070596 0.6526621 0.6203062 0.6159569 0.6234357
#> [22] 0.5851770 0.5831123 0.9411752 0.9192082 0.8158554 0.6788430 0.7118105
#> [29] 0.6767945 0.8694307 0.8247718 0.6654835 0.6769155 0.6751160 0.6136048
#> [36] 0.8526010 0.8431762 0.7110224 0.7084562 0.7209227 0.7191071 0.6533683
#> [43] 0.7629557 0.7546727 0.7084879 0.6982745 0.6889437 0.6796252 0.6456600
#> [50] 1.1001602 0.7741629 0.7345502 0.7066990 0.6958463 0.9311880 0.8558849
#> [57] 0.8467353 0.7972344 0.7669733 0.7844239 0.7410834 0.7013396 0.8389582
#> [64] 0.7946941 0.8067596 0.7511686 0.7470219 0.6840763 1.0047320 0.8978433
#> [71] 0.8398436 0.7544793 0.7272156 0.8305171 0.7305346 0.7315924 0.7146458
#> [78] 0.8890375 0.7079133 0.6920796 0.7008547 0.6878918 0.7263276 0.7144593
#> [85] 0.6845575 0.6684377 0.6696673 0.7001397 0.6683025 0.8476000
#> attr(,"label")
#> [1] "Predicted values"
( p <- predict(gm, test[,c(x, "DISTANCE")], back.transform = "Naihua") )
#> Making population level predictions
#> Naihua back-transformation using: 
#> 	  y ~ exp(y-hat) regression with no intercept,
#> 	  does not assume normally distributed errors
#>  [1] 1.0073581 1.0038246 0.9350617 0.8973507 0.9116684 0.7956627 0.9662661
#>  [8] 0.9479500 0.9340152 0.8748092 0.8549898 0.8100380 0.8330870 0.7991371
#> [15] 0.9766359 0.8834735 0.8701167 0.8031744 0.7633569 0.7580045 0.7672080
#> [22] 0.7201264 0.7175855 1.1582224 1.1311895 1.0040022 0.8353929 0.8759632
#> [29] 0.8328721 1.0699327 1.0149748 0.8189526 0.8330209 0.8308064 0.7551099
#> [36] 1.0492218 1.0376236 0.8749934 0.8718353 0.8871768 0.8849425 0.8040435
#> [43] 0.9389032 0.9287099 0.8718744 0.8593056 0.8478230 0.8363555 0.7945576
#> [50] 1.3538714 0.9526948 0.9039470 0.8696730 0.8563175 1.1459320 1.0532630
#> [57] 1.0420034 0.9810869 0.9438472 0.9653221 0.9119869 0.8630776 1.0324328
#> [64] 0.9779608 0.9928088 0.9243977 0.9192948 0.8418331 1.2364362 1.1048976
#> [71] 1.0335223 0.9284720 0.8949209 1.0220451 0.8990053 0.9003070 0.8794523
#> [78] 1.0940611 0.8711673 0.8516821 0.8624808 0.8465286 0.8938281 0.8792228
#> [85] 0.8424253 0.8225880 0.8241012 0.8616010 0.8224217 1.0430675
#> attr(,"label")
#> [1] "Predicted values"

 
# Using grouped data
test <- nlme::groupedData(stats::as.formula(paste(paste("DPS", 1, sep = " ~ "), 
          "FROM_SITE", sep = " | ")), 
      data = test[,c("DPS", "FROM_SITE", x, "DISTANCE")])

( p <- predict(gm, test, groups = "FROM_SITE") )
#> Making group level (individual) predictions
#> Warning: the condition has length > 1 and only the first element will be used
#>           1           1           1           1           1           1 
#> -0.23580292 -0.23931678 -0.31027689 -0.35144263 -0.33561304 -0.47171411 
#>           2           2           2           2           2           2 
#> -0.26035692 -0.27949437 -0.29430343 -0.35979039 -0.38270662 -0.43671496 
#>           2           2           3           3           3           3 
#> -0.40865812 -0.45026364 -0.28026281 -0.38051545 -0.39574934 -0.47580477 
#>           3           3           3           3           3           4 
#> -0.52665108 -0.53368738 -0.52161874 -0.58495000 -0.58848455 -0.08326620 
#>           4           4           4           4           4           5 
#> -0.10688291 -0.22615838 -0.41000568 -0.36258381 -0.41302777 -0.19139403 
#>           5           5           5           5           5           6 
#> -0.24412594 -0.45871879 -0.44168627 -0.44434819 -0.53988167 -0.19471307 
#>           6           6           6           6           6           6 
#> -0.20582871 -0.37630072 -0.37991650 -0.36247276 -0.36499442 -0.46086369 
#>           7           7           7           7           7           7 
#> -0.31953583 -0.33045173 -0.39360286 -0.40812358 -0.42157624 -0.43519436 
#>           7           8           8           8           8           8 
#> -0.48646275  0.02815312 -0.32327574 -0.37579964 -0.41445311 -0.42992916 
#>           9           9           9           9           9           9 
#> -0.14882454 -0.23314982 -0.24389762 -0.30413697 -0.34283373 -0.32033618 
#>           9           9          10          10          10          10 
#> -0.37717248 -0.43229342 -0.23965765 -0.29386129 -0.27879277 -0.35018844 
#>          10          10          11          11          11          11 
#> -0.35572396 -0.44374903 -0.05856053 -0.17104114 -0.23782105 -0.34500885 
#>          11          12          12          12          12          13 
#> -0.38181377 -0.28437744 -0.41264943 -0.41120251 -0.43463899 -0.14381295 
#>          13          14          14          14          14          15 
#> -0.37163075 -0.45239099 -0.43979142 -0.45846035 -0.40409086 -0.34431282 
#>          15          16          16          17          18          19 
#> -0.38706615 -0.51009474 -0.50825694 -0.41007320 -0.44716179          NA 
#> attr(,"label")
#> [1] "Predicted values"
  rmse(back.transform(p), back.transform(ralu.model[,"DPS"][-sidx]))
#> [1] NA

# Specify unconstrained gravity model (generally, not recommended)  
( gm <- gravity(y = "DPS", x = x, d = "DISTANCE", group = "FROM_SITE", 
                data = train, ln = FALSE, constrained=TRUE) )
#> [1] "Running singly-constrained gravity model"
#> Gravity model
#> 
#> Linear mixed-effects model fit by REML
#>   Data: gdata 
#>        AIC       BIC logLik
#>   -136.514 -113.7206 77.257
#> 
#> Random effects:
#>  Formula: DPS ~ 1 | FROM_SITE
#>         (Intercept)   Residual
#> StdDev:  0.04331721 0.09251116
#> 
#> Fixed effects:  list(fmla) 
#>                 Value Std.Error DF   t-value p-value
#> (Intercept) -5.021556 2.2138316 79 -2.268265  0.0260
#> DEPTH_F      0.016064 0.0142664 14  1.126013  0.2791
#> HLI_F        0.057311 0.2058492 14  0.278415  0.7848
#> CTI_F       -0.056346 0.0637005 14 -0.884544  0.3913
#> cti         -0.168146 0.0981068 79 -1.713907  0.0905
#> ffp          0.897728 0.2040685 79  4.399147  0.0000
#> DISTANCE    -0.125777 0.0154870 79 -8.121503  0.0000
#>  Correlation: 
#>          (Intr) DEPTH_ HLI_F  CTI_F  cti    ffp   
#> DEPTH_F  -0.626                                   
#> HLI_F    -0.824  0.859                            
#> CTI_F     0.330 -0.656 -0.560                     
#> cti      -0.024  0.050  0.012 -0.128              
#> ffp      -0.777  0.129  0.301  0.003 -0.138       
#> DISTANCE -0.275  0.107  0.140 -0.050  0.464  0.164
#> 
#> Standardized Within-Group Residuals:
#>         Min          Q1         Med          Q3         Max 
#> -2.96880469 -0.49096445  0.06522563  0.53632544  1.79390405 
#> 
#> Number of Observations: 100
#> Number of Groups: 18 

( p <- predict(gm, test[,c(x, "DISTANCE")]) )
#> Making population level predictions
#> Warning: the condition has length > 1 and only the first element will be used
#>  [1] -0.25329785 -0.25681171 -0.32777182 -0.36893756 -0.35310797 -0.48920904
#>  [7] -0.29494511 -0.31408256 -0.32889162 -0.39437858 -0.41729481 -0.47130315
#> [13] -0.44324631 -0.48485183 -0.28427044 -0.38452308 -0.39975697 -0.47981240
#> [19] -0.53065871 -0.53769501 -0.52562637 -0.58895763 -0.59249218 -0.11374265
#> [25] -0.13735936 -0.25663483 -0.44048213 -0.39306026 -0.44350423 -0.19303334
#> [31] -0.24576525 -0.46035810 -0.44332558 -0.44598750 -0.54152098 -0.21258035
#> [37] -0.22369599 -0.39416800 -0.39778379 -0.38034004 -0.38286170 -0.47873097
#> [43] -0.32367198 -0.33458788 -0.39773901 -0.41225972 -0.42571239 -0.43933051
#> [49] -0.49059890  0.04233915 -0.30908972 -0.36161362 -0.40026709 -0.41574314
#> [55] -0.12441081 -0.20873610 -0.21948389 -0.27972325 -0.31842001 -0.29592246
#> [61] -0.35275876 -0.40787969 -0.22871113 -0.28291477 -0.26784625 -0.33924191
#> [67] -0.34477744 -0.43280251 -0.04839582 -0.16087642 -0.22765633 -0.33484413
#> [73] -0.37164905 -0.23882341 -0.36709540 -0.36564847 -0.38908496 -0.17073254
#> [79] -0.39855034 -0.42117103 -0.40857146 -0.42724038 -0.37287089 -0.38934601
#> [85] -0.43209934 -0.45592882 -0.45409103 -0.40959210 -0.45613108 -0.21846319
#> attr(,"label")
#> [1] "Predicted values"
  rmse(back.transform(p), back.transform(ralu.model[,"DPS"][-sidx])) 
#> [1] 0.0761695